better seperation of logic
