+++ webapp/src/components/views/HomeView.tsx
@@ -1,7 +1,7 @@
 import NavigationButtons from "./NavigationButtons";
 import ListView from "./List";
 import MapErrorView from "./MapError";
-import MapView from "./Map";
+import MapView from "./MapView";
 import SearchBoxView from "./SearchBoxView";
 
 import { Places } from "../../../types/types";
@@ -26,6 +26,9 @@ type HomeViewProps = {
   searchRequested: boolean;
   setSearchRequested: React.Dispatch<React.SetStateAction<boolean>>;
   triggerSearch: () => void;
+
+
+  mapRef: React.RefObject<HTMLDivElement>;
 };
 
 export default function HomeView({ 
@@ -38,7 +41,9 @@ export default function HomeView({
 
   searchRequested,
   setSearchRequested,
-  triggerSearch
+  triggerSearch,
+
+  mapRef
 
 }: HomeViewProps) {
 
@@ -57,7 +62,7 @@ export default function HomeView({
           <ListView></ListView>
         </div>
         <div className="mainGridTwo">
-          <MapView mapsapiKey = {apiGMapsKey} placesAPIKey={apiGPlaceskey} origin={origin} setPlaces={setPlaces} ></MapView>
+          <MapView mapRef={mapRef} ></MapView>
 
           {/* <MapView
             mapsapiKey={apiGMapsKey}
